Update GettingStarted.tid
I've just started using Tiddly, and even though it's a great tool, I almost gave up because TiddlyFox would not work. After much persistence and a whole hour searching for the problem, the solution was a simple uncheck on "read-only" of the empty-file.html attributes. I bet this will happen to another Windows user, and this little tip on "Troubleshooting" section of the "Getting started" could save a lot of headaches.